Chapter 2. 추상화의 개념적 결합 분리(#3)
