+++ src/main/java/com/objects/objects/movieReservation/Movie.java
@@ -1,6 +1,6 @@
 package com.objects.objects.movieReservation;
 
-import com.objects.objects.movieReservation.discountPolicy.DiscountPolicy;
+import com.objects.objects.movieReservation.discountPolicy.DefaultDiscountPolicy;
 
 import java.time.Duration;
 
@@ -12,9 +12,9 @@ public class Movie {
     private String title;
     private Duration runningTime;
     private Money fee;
-    private DiscountPolicy discountPolicy;
+    private DefaultDiscountPolicy discountPolicy;
 
-    public Movie(String title, Duration runningTime, Money fee, DiscountPolicy discountPolicy) {
+    public Movie(String title, Duration runningTime, Money fee, DefaultDiscountPolicy discountPolicy) {
         this.title = title;
         this.runningTime = runningTime;
         this.fee = fee;
