+++ src/main/java/com/objects/objects/movieReservation/discountPolicy/DiscountPolicy.java
@@ -2,33 +2,10 @@ package com.objects.objects.movieReservation.discountPolicy;
 
 import com.objects.objects.movieReservation.Money;
 import com.objects.objects.movieReservation.Screening;
-import com.objects.objects.movieReservation.discountCondition.DiscountCondition;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 /**
- * 금액 할인 정책과 비율 할인 정책
- * - 계산 방식에만 차이가 있으므로 추상 클래스로 생성
- * - Template Method Pattern
+ * 추상화 수준에서의 개념적 결합 방지를 위한 인터페이스 분리
  */
-public abstract class DiscountPolicy {
-    private List<DiscountCondition> conditions = new ArrayList<>();
-
-    public DiscountPolicy(DiscountCondition... conditions) {
-        this.conditions = Arrays.asList(conditions);
-    }
-
-    public Money calculateDiscountAmount(Screening screening) {
-        for (DiscountCondition each : conditions) {
-            if (each.isSatisfiedBy(screening)) {
-                return getDiscountAmount(screening);
-            }
-        }
-
-        return Money.ZERO;
-    }
-
-    abstract protected Money getDiscountAmount(Screening screening);
+public interface DiscountPolicy {
+    Money calculateDiscountAmount(Screening screening);
 }
