+++ src/main/java/com/objects/objects/movieReservation/discountPolicy/NoneDiscountPolicy.java
@@ -3,9 +3,9 @@ package com.objects.objects.movieReservation.discountPolicy;
 import com.objects.objects.movieReservation.Money;
 import com.objects.objects.movieReservation.Screening;
 
-public class NoneDiscountPolicy extends DiscountPolicy {
+public class NoneDiscountPolicy implements DiscountPolicy {
     @Override
-    protected Money getDiscountAmount(Screening screening) {
+    public Money calculateDiscountAmount(Screening screening) {
         return Money.ZERO;
     }
 }
