+++ src/main/java/com/objects/objects/movieReservation/discountPolicy/PercentDiscountPolicy.java
@@ -7,7 +7,7 @@ import com.objects.objects.movieReservation.discountCondition.DiscountCondition;
 /**
  * 비율 할인 정책
  */
-public class PercentDiscountPolicy extends DiscountPolicy {
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
     private double percent;
 
     public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
